+++ src/tests/Interop/COM/NativeServer/AggregationTesting.h
@@ -98,12 +98,12 @@ private:
             return _implOuter->QueryInterface(riid, ppvObject);
         }
 
-        STDMETHODIMP_(ULONG) AddRef(void) 
+        STDMETHODIMP_(uint32_t) AddRef(void)
         {
             return _implOuter->AddRef();
         }
 
-        STDMETHODIMP_(ULONG) Release(void) 
+        STDMETHODIMP_(uint32_t) Release(void)
         {
             return _implOuter->Release();
         }
